The build system changes required to move actor headers to .objs/
But this doesn't work because...

A file in fdbclient/ would have

 -I . -I fdbclient -I .objs -I .objs/fdbclient

Suppose fdbclient/foobar.cpp has #include "foobar.h"
...which has a #include "flow/IndexedSet.h"
...which has a #include "IndexedSet.actor.h"
...which has a #include "IndexedSet.actor.g.h"

IndexedSet.h can find IndexedSet.actor.h, because it's in the same folder.  But
IndexedSet.actor.g.h isn't in the same folder, and we don't know that we should
also have a `-I .objs/flow` on the command line, so we can't find the file.